+++ src/App.tsx
@@ -4,60 +4,13 @@ import type { Movie } from "./types/Movie";
 import Header from "./components/Header";
 import MovieForm from "./components/MovieForm";
 import MovieList from "./components/MovieList";
+import { useMovies } from "./hooks/useMovies";
 
 function App() {
   const [themeName, setThemeName] = useState<ThemeName>("light");
-  const [movies, setMovies] = useState<Movie[]>([]);
-  const [error, setError] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const { movies, isLoading, error, addMovie } = useMovies();
 
   const currentTheme = themes[themeName];
-
-  useEffect(() => {
-    const fetchMovies = async () => {
-      setIsLoading(true);
-      try {
-        const res = await fetch("/api/movies");
-        if (!res.ok) throw new Error("영화 데이터를 불러오지 못했습니다.");
-        const data: Movie[] = await res.json();
-        setMovies(data);
-      } catch (err: unknown) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError(String(err));
-        }
-      } finally {
-        setIsLoading(false);
-      }
-    };
-    fetchMovies();
-  }, []);
-
-  const handleAddMovie = async (movie: Omit<Movie, "id">) => {
-    try {
-      setIsLoading(true);
-      const res = await fetch("/api/movies", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(movie),
-      });
-
-      if (!res.ok) throw new Error("영화를 추가하지 못했습니다.");
-
-      const savedMovie: Movie = await res.json();
-      setMovies((prev) => [...prev, savedMovie]);
-      setError(null);
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        setError(err.message);
-      } else {
-        setError(String(err));
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
 
   return (
     <div
@@ -83,7 +36,7 @@ function App() {
           {error}
         </div>
       )}
-      <MovieForm themeName={themeName} onAddMovie={handleAddMovie} />
+      <MovieForm themeName={themeName} onAddMovie={addMovie} />
       <MovieList movies={movies} isLoading={isLoading} themeName={themeName} />
     </div>
   );
